+++ src/scene_cluster_lod.cpp
@@ -290,9 +290,11 @@ void Scene::buildGeometryClusterLodMeshoptimizer(const ProcessingInfo& processin
   clodConfig clodInfo = m_config.meshoptPreferRayTracing ? clodDefaultConfigRT(m_config.clusterTriangles) :
                                                            clodDefaultConfig(m_config.clusterTriangles);
 
-  clodInfo.max_vertices      = m_config.clusterVertices;
-  clodInfo.partition_size    = m_config.clusterGroupSize;
-  clodInfo.partition_spatial = true;
+  clodInfo.cluster_fill_weight  = m_config.meshoptFillWeight;
+  clodInfo.cluster_split_factor = m_config.meshoptSplitFactor;
+  clodInfo.max_vertices         = m_config.clusterVertices;
+  clodInfo.partition_size       = m_config.clusterGroupSize;
+  clodInfo.partition_spatial    = true;
 
   // account for meshopt_partitionClusters's using a target value with a higher worst case
   while((clodInfo.partition_size + clodInfo.partition_size / 3) > m_config.clusterGroupSize)